pyproject.toml
[tool.poetry]
name = "drf-course-work"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
django-dump-load-utf8 = "^0.0.4"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
coverage = "^7.5.1"
django-celery-beat = "^2.6.0"
redis = "^5.0.4"
eventlet = "^0.36.1"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
setuptools = "^69.5.1"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = '.venv'
warn_unreachable = true
ignore_missing_imports = true

[tool.black]
line-length = 119

[tool.isort]
line_length = 119